started splitting out a stand alone resource viewer instead of mixing it with log tab. seems to cause small issue on first few requests at ui layer only because web browser still gets data but ui shows empty content

+++ src/com/thedeanda/ajaxproxy/ui/FileTrackerTableModel.java
@@ -62,7 +62,7 @@ public class FileTrackerTableModel extends AbstractTableModel {
 			return files.get(row);
 	}
 
-	public void trackFile(String url, int duration) {
+	public void trackFile(String url, long duration) {
 		int indexAdded = -1;
 		int indexMod = -1;
 		synchronized (files) {
